# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PlanteLand API",
        "description": "Arrosage pas automatique",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/api/plantes": {
            "get": {
                "summary": "Récupérer toutes les plantes",
                "operationId": "d1176f9e2d48bf4288712f2836753ff9",
                "responses": {
                    "200": {
                        "description": "Liste de plantes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Plantes"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Plante"
                ],
                "summary": "Create a new plante",
                "description": "Creates a new plante with the given data",
                "operationId": "68706de9267b956aa7c8cf0c8d81d0ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "common_name",
                                    "watering_general_benchmark",
                                    "watering_general_benchmark.value",
                                    "watering_general_benchmark.unit"
                                ],
                                "properties": {
                                    "common_name": {
                                        "type": "string",
                                        "example": "Rose"
                                    },
                                    "watering_general_benchmark": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "value": {
                                                    "type": "number",
                                                    "example": 1.5
                                                },
                                                "unit": {
                                                    "type": "string",
                                                    "example": "liters"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plante created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante créée avec succès"
                                        },
                                        "plante": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    }
                }
            }
        },
        "/api/plantes/{id}": {
            "get": {
                "tags": [
                    "Plante"
                ],
                "summary": "Get a specific plante by ID",
                "description": "Returns a single plante by its ID",
                "operationId": "af89581451fdba20804c3a87337d9286",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plante",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The requested plante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "common_name": {
                                            "type": "string",
                                            "example": "Rose"
                                        },
                                        "city": {
                                            "type": "string",
                                            "example": "Paris"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plante not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Plante"
                ],
                "summary": "Delete a plante by ID",
                "description": "Deletes a plante with the given ID",
                "operationId": "73c6f5590f18239a3b4cc27527ca366f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the plante to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plante deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plante supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plante not found"
                    }
                }
            }
        },
...
        "/api/plante/plantes": {
            "get": {
                "summary": "Récupérer les plantes par utilisateur",
                "operationId": "de30af9729f60a6d58898650ac64ab88",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des plantes de l'utilisateur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Liste des plantes de l'utilisateur"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Utilisateur non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Plantes": {
                "title": "Plantes",
                "description": "Model representing a plante",
                "properties": {
                    "id": {
                        "description": "Unique identifier for the plante",
                        "type": "integer",
                        "example": 1
                    },
                    "common_name": {
                        "description": "The common name of the plante",
                        "type": "string",
                        "example": "Rose"
                    },
                    "watering_general_benchmark": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "value": {
                                    "type": "number",
                                    "example": 1.5
                                },
                                "unit": {
                                    "type": "string",
                                    "example": "liters"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "city": {
                        "description": "The city where the plante is located",
                        "type": "string",
                        "example": "Paris"
                    },
                    "": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            },
            "User": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "description": "User ID",
                        "type": "integer"
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string"
                    },
                    "password": {
                        "description": "User's password",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "Timestamp when email was verified",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Account creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last account update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "BearerAuth": []
        }
    ]
}